audit_iq_echo_rag/pyproject.toml
[project]
name = "audit_iq_echo_rag"
version = "0.1.0"
description = "AuditIQ ECHO RAG Tool - GT Guidelines and Policy search using Azure Cognitive Search"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.186.1",
    "azure-search-documents>=11.4.0",
    "azure-core>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "tool"